+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     (Pedro Carriço and Erik Ostrom) [Issue #432]
 *   Raise exception on PhantomJS "status: fail" result (i.e DNS issue) instead
     of returning minimal HTML body (Dean Holdren) [Issue #473]
+*   Render full window size when document has no height (Kevin McConnell)
 
 ### 1.5.0 ###